requirements.txt
Flask
flask-cors
deepface
numpy
Pillow
opencv-python-headless